+++ x/transfer/module.go
@@ -8,7 +8,7 @@ import (
 	"github.com/cosmos/ibc-go/v3/modules/apps/transfer/keeper"
 	"github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
-	"github.com/neutron-org/neutron/x/sudo"
+	"github.com/neutron-org/neutron/internal/sudo"
 )
 
 type IBCModule struct {
